Bugfix/Light sensors adc has been revoked and is send as a float/string instead of integer

+++ src/sensor_manager.h
@@ -6,7 +6,7 @@ typedef struct {
     uint8_t temp_d;
     uint8_t hum_i;
     uint8_t hum_d;
-    uint16_t light;
+    float light;
     float rain;
     float wind_speed;
     uint16_t wind_dir;
